requirements.txt
# Core Dependencies
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0

# MCP Server Dependencies (for AI-native automation)
mcp>=1.26.0
uvicorn>=0.40.0
starlette>=0.52.0
pydantic>=2.11.0
pydantic-settings>=2.12.0
httpx>=0.28.0
httpx-sse>=0.4.3
